Fix _checkbw processing
git-svn-id: svn://svn.mplayerhq.hu/rtmpdump@46 400ebc74-4327-4243-bc38-086b20814532

+++ rtmp.h
@@ -175,7 +175,7 @@ class CRTMP
       bool SendConnectPacket();
       bool SendServerBW();
       bool SendCheckBW();
-      bool SendCheckBWResult(int txn);
+      bool SendCheckBWResult(double txn);
       bool SendCtrl(short nType, unsigned int nObject, unsigned int nTime = 0);
       bool SendBGHasStream(double dId, char *playpath);
       bool SendCreateStream(double dStreamId);
